Contributer: HaskellChad. Added functionality for saving the models and loading them, later I will work on having it generate samples based on a user's preference. The CLI interface has also been improved, but there is a warning that it is giving me when saving the model and I have no idea why (it did this beforehand too, its probably nothing).
